requirements.txt
# Core ML stack
torch>=2.2.0
transformers>=4.47.0
datasets>=3.0.0
accelerate>=1.2.0
peft>=0.13.0
trl>=0.12.0
unsloth>=2025.1.0
bitsandbytes>=0.45.0

# Utilities
pyyaml>=6.0.2
numpy>=2.1.0
rich>=13.9.4
wandb>=0.19.0
tqdm>=4.67.0

# Optional integrations
mistralai>=1.5.0
gradio>=5.9.0
